Export Map and Directory Fuzz Models
