+++ packages/dds/map/src/test/tsconfig.json
@@ -7,6 +7,8 @@
 		"noUnusedLocals": false, // Need it so memory tests can declare local variables just for the sake of keeping things in memory
 		"noUncheckedIndexedAccess": false,
 		"exactOptionalPropertyTypes": false,
+		"declaration": true,
+		"declarationMap": true,
 	},
 	"include": ["./**/*"],
 	"references": [
